requirements.txt
flask
gunicorn
PyPDF2
Pillow
pdf2image